+++ src/components/EditionDetail/Edition.test.tsx
@@ -75,7 +75,7 @@ describe("Renders edition component when given valid edition", () => {
       screen
         .getAllByText("Copyright: Public Domain when viewed in the US")[0]
         .closest("a").href
-    ).toContain("/license");
+    ).toContain("/copyright");
   });
 
   test("Featured Card, which has publisher 'Miller', shows up once", () => {